pyproject.toml
[tool.poetry]
name = "sensors2mqtt"
version = "0.0.1"
authors = [
  "Dries Claerbout <claerbout@hotmail.com>" ,
]
homepage  = "https://github.com/DCL777/sensors2mqtt"
repository   = "https://github.com/DCL777/sensors2mqtt"
description = "Universal Sensor data transporter over MQTT"
keywords = ["MQTT", "Sensors", "DS18B20", "INA219", "INA226", "PMS5003", "SystemInfo", "Dahua", "Pulse Counter", "Ultrasonic", "JSN-SR04T","JSN-SR04T-2.0","JSN-SR04T-3.0","HC-SR04","HY-SRF05"]
readme = "README.md"
license = "GPL-3.0-only"
classifiers = [
    "Natural Language :: English",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3.0",
    "Operating System :: POSIX :: Linux",
	"Development Status :: 4 - Beta",

]


[tool.poetry.dependencies]
certifi = "^2022.6.15"
python = "^3.7"
idna = "^3.3"
python-dateutil = "^2.8.2"
paho-mqtt = "^1.6.1"
urllib3  = "^1.26.10"
pyyaml = "^6.0"
smbus2 = "^0.4.2"
psutil = "^5.9.1"
plantower = "^0.0.11"
setuptools = "^63.2.0"
libssl-dev = "^1.1.0"
pycurl = "^7.45.1"
distro = "^1.7.0"


[tool.poetry.dev-dependencies]


[tool.poetry.urls]
"Bug Tracker" = "https://github.com/DCL777/sensors2mqtt/issues"
Homepage = "https://github.com/DCL777/sensors2mqtt"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
